% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{add_noisevector}
\alias{add_noisevector}
\title{Add Noise-vector to Sparse Matrix}
\usage{
add_noisevector(x, noise_vector)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix`.}

\item{noise_vector}{A numeric vector of `length(nrow(x))`.}
}
\value{
A sparse matrix with added noise.
}
\description{
Adds a noise-vector to a sparse matrix.
}
\examples{
\dontrun{
# Assuming `sparse_mat` is a dgCMatrix
NV = abs(rnorm(100,mean(as.matrix(x)),sd(as.matrix(x))))
add_noisevector(sparse_mat[1:100,], noise_vector = NV)
}

}
