% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate Correlation per cell to Ground Truth}
\usage{
calculate_correlation(
  x,
  grouping_vector,
  bin_size,
  ground_truth,
  method = "pearson"
)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix` with cells to correlate.}

\item{grouping_vector}{A vector indicating the group (cell type) for each column in `x`.}

\item{bin_size}{An integer specifying the bin size in base pairs.}

\item{ground_truth}{A sparse matrix containing the ground truth pseudobulk per cell type.}

\item{method}{A character string specifying the correlation method, either 'pearson' or 'spearman'; defaults to 'pearson'.}
}
\value{
A data frame with summary statistics of correlation values for each cell type, including
  \describe{
    \item{group}{The cell type or "complete" for overall statistics.}
    \item{cor_method}{The correlation method used.}
    \item{mu}{Mean correlation value.}
    \item{sd}{Standard deviation of the correlation values.}
    \item{N}{Number of cells in the group.}
  }
}
\description{
This function calculates the correlation between each cell in a matrix and a ground truth pseudobulk,
grouped by cell type, using a specified correlation method.
}
\examples{
data(zeller_H3K4me3_matrix)
data(zeller_H3K4me3_metadata)

ground_truth_mat <- generate_ground_truth(
  x = zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata,
  bin_size = 50000,
  fzc_sd = 0)

correlation_stats <- calculate_correlation(zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata,
  bin_size = 50000,
  ground_truth = ground_truth_mat,
  method = "pearson")

}
