% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{calculate_fzc}
\alias{calculate_fzc}
\title{Calculate Fraction of Zero-Bins}
\usage{
calculate_fzc(x)
}
\arguments{
\item{x}{A numeric vector for which the fraction of zero elements is calculated.}
}
\value{
A numeric value between 0 and 1 representing the fraction of zero values in `x`.
}
\description{
Computes the fraction of elements in a numeric vector that are zero.
}
\examples{
\dontrun{
calculate_fzc(c(0, 1, 0, 2, 3))
calculate_fzc(c(0, 0, 0, 0))
calculate_fzc(c(1, 2, 3, 4))
}

}
