% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{calculate_pseudobulk}
\alias{calculate_pseudobulk}
\title{Calculate Pseudobulk from Sparse Matrix}
\usage{
calculate_pseudobulk(x, grouping_vector = NULL)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix`.}

\item{grouping_vector}{A vector indicating group assignments for each column in `x`. If `NULL`,
all columns will be grouped into a single pseudobulk profile.}
}
\value{
A sparse matrix of class `dgCMatrix` where each column represents the pseudobulk profile
of a group specified in `grouping_vector`.
}
\description{
Aggregates columns of a sparse matrix into pseudobulk profiles based on a grouping vector.
}
\examples{
\dontrun{
# Assuming `sparse_mat` is a dgCMatrix and `groups` is a vector of group labels
calculate_pseudobulk(sparse_mat, grouping_vector = groups)
}
}
