% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{calculate_simic}
\alias{calculate_simic}
\title{Calculate SIMIC: Similarity Index for Clustering}
\usage{
calculate_simic(
  x,
  grouping_vector,
  ab_initio = NULL,
  HVB_q = 0.95,
  K = 10,
  N_cores = 2
)
}
\arguments{
\item{x}{A sparse matrix of data for similarity calculation.}

\item{grouping_vector}{A vector indicating group assignments for each cell in `x`.}

\item{ab_initio}{Optional precomputed similarity matrix. If `NULL`, it will be calculated.}

\item{HVB_q}{Quantile of most variable bins to consider (e.g., 0.95 means top-5 percent variable bins).}

\item{K}{Integer, number of nearest neighbors to consider in calculating similarity.}

\item{N_cores}{Integer, number of cores for parallel processing.}
}
\value{
A data frame with observed similarity, expected similarity, and their ratio (delta).
}
\description{
This function calculates the SIMIC score, which compares observed similarity to expected similarity
for clustering cell types. The user can specify quantiles for high-variance bins (HVB) to focus on
the most variable features.
}
\examples{
data(zeller_H3K4me3_matrix)
data(zeller_H3K4me3_metadata)

sim_mat <- generate_ab_initio(
  x = zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata)

simicstats <- calculate_simic(
  x = zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata,
  ab_initio = sim_mat,
  HVB_q = 0.95,
  K = 10,
  N_cores = 2)

}
