% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{filter_by_fzc}
\alias{filter_by_fzc}
\title{Filter Matrix by Fraction of Zero Counts (FZC)}
\usage{
filter_by_fzc(x, sd_magnitude = 2)
}
\arguments{
\item{x}{A sparse matrix of class `Matrix` or `dgCMatrix`.}

\item{sd_magnitude}{A numeric value indicating the number of standard deviations below the mean FZC
used to set the threshold. Default is 2.}
}
\value{
A sparse matrix containing only the columns with an FZC above the calculated threshold.
}
\description{
Filters columns of a sparse matrix based on the fraction of zero values in each column.
Columns with an FZC below a specified threshold (mean FZC minus `sd_magnitude` standard deviations)
are excluded from the returned matrix.
}
\examples{
\dontrun{
# Assuming `sparse_mat` is a dgCMatrix
filter_by_fzc(sparse_mat, sd_magnitude = 2)
}

}
