% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{generate_ab_initio}
\alias{generate_ab_initio}
\title{Generate Similarity Matrix Using Cosine Distance}
\usage{
generate_ab_initio(x, grouping_vector)
}
\arguments{
\item{x}{A sparse matrix containing data to be grouped by the `grouping_vector`.}

\item{grouping_vector}{A vector specifying group assignments for each column (cell) in `x`.}
}
\value{
A similarity matrix calculated using cosine distance on pseudobulk profiles.
}
\description{
This function calculates the cosine similarity matrix for pseudobulks.
}
\examples{
data(zeller_H3K4me3_matrix)
data(zeller_H3K4me3_metadata)

sim_mat <- generate_ab_initio(
  x = zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata)

}
