% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{generate_ground_truth}
\alias{generate_ground_truth}
\title{Generate Ground Truth Pseudobulks from Sparse Matrix}
\usage{
generate_ground_truth(x, grouping_vector, bin_size, min_RD = 0, fzc_sd = NULL)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix`.}

\item{grouping_vector}{A vector indicating the grouping for each column in `x`.}

\item{bin_size}{An integer specifying the expected bin size in base pairs.}

\item{min_RD}{An integer specifying the minimum read depth required for inclusion.}

\item{fzc_sd}{An optional numeric value to filter by fraction of zero bins (FZC).}
}
\value{
A sparse matrix of class `dgCMatrix` containing the ground truth pseudobulks.
}
\description{
This function processes a sparse matrix to generate ground truth pseudobulks,
filtering by sparsity and minimum read depth. It expands the matrix to include
all expected genomic bins and calculates pseudobulks based on a provided grouping vector.
}
\examples{
data(zeller_H3K4me3_matrix)
data(zeller_H3K4me3_metadata)

zeller_ground_truth <- generate_ground_truth(
  zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata,
  bin_size = 50000,
  min_RD = 10
  )

}
