% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{generate_in_silico}
\alias{generate_in_silico}
\title{Generate In Silico Datasets with Noise and Downsampling}
\usage{
generate_in_silico(
  x,
  grouping_vector,
  bin_size,
  target_RD,
  noise_level = 0,
  noise_vector = NULL,
  min_RD = 0,
  fzc_sd = NULL
)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix`.}

\item{grouping_vector}{A vector indicating the grouping for each column in `x`.}

\item{bin_size}{An integer specifying the expected bin size in base pairs.}

\item{target_RD}{A numeric value specifying the target read depth for downsampling.}

\item{noise_level}{A numeric value indicating the noise level to be added; defaults to 0.}

\item{noise_vector}{If `noise_level` is set to 0, the user can add a vector containing user-specified noise-values per bin.}

\item{min_RD}{An integer specifying the minimum read depth required for inclusion.}

\item{fzc_sd}{An optional numeric value for filtering by fraction of zero bins (FZC).}
}
\value{
A list with two elements:
  \describe{
    \item{ground_truth}{A sparse matrix of class `dgCMatrix` containing the ground truth pseudobulks.}
    \item{in_silico_matrix}{A sparse matrix with added noise and downsampling, representing the in silico dataset.}
  }
}
\description{
This function generates an in silico dataset by creating pseudobulks from a sparse matrix,
optionally adding noise and downsampling to achieve a target read depth.
}
\examples{
data(zeller_H3K4me3_matrix)
data(zeller_H3K4me3_metadata)

in_silico_output <- generate_in_silico(zeller_H3K4me3_matrix,
  grouping_vector = zeller_H3K4me3_metadata,
  bin_size = 50000,
  target_RD = 1000,
  noise_level = 0.1)

}
