% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_data.R
\name{increase_noise}
\alias{increase_noise}
\title{Add Noise to Sparse Matrix}
\usage{
increase_noise(x, increase_factor = 0)
}
\arguments{
\item{x}{A sparse matrix of class `dgCMatrix`.}

\item{increase_factor}{A numeric value indicating the factor by which to increase the original noise level. Default is 0.}
}
\value{
A sparse matrix with added noise.
}
\description{
Adds noise to a sparse matrix by increasing the original noise level by a specified factor.
}
\examples{
\dontrun{
# Assuming `sparse_mat` is a dgCMatrix
increase_noise(sparse_mat, increase_factor = 2)
}

}
