% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{minmax}
\alias{minmax}
\title{Min-Max Scaling}
\usage{
minmax(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector to be rescaled.}

\item{na.rm}{A logical value indicating whether `NA` values should be removed before scaling.
Defaults to `TRUE`.}
}
\value{
A numeric vector with values scaled to the [0, 1] range.
}
\description{
This function applies min-max scaling to a numeric vector, rescaling values to a [0, 1] range.
}
\examples{
minmax(c(1, 2, 3, 4, 5))
minmax(c(1, NA, 3, 4, 5), na.rm = TRUE)

}
