% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeller_H3K4me3.R
\docType{data}
\name{zeller_H3K4me3_matrix}
\alias{zeller_H3K4me3_matrix}
\title{H3K4me3 sortChIC data}
\format{
A bins by cells sparse matrix
}
\usage{
zeller_H3K4me3_matrix
}
\description{
These data sets contain a subset of the cells from the H3K4me3 dataset of Zeller et al. (2023)
}
\details{
Contains a sparse matrix of 100 B-cells and 100 Erythrocytes, binned at 50kb.
}
\examples{
test_matrix <- zeller_H3K4me3_matrix
test_matrix_clusters <- zeller_H3K4me3_metadata
}
\keyword{datasets}
