% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeller_H3K4me3.R
\docType{data}
\name{zeller_H3K4me3_peaks}
\alias{zeller_H3K4me3_peaks}
\title{H3K4me3 sortChIC peaks}
\format{
A `GRangesList` object with two entries
}
\usage{
zeller_H3K4me3_peaks
}
\description{
These datasets contain filtered peaks for two cell types from the H3K4me3 dataset of Zeller et al. (2023)
}
\details{
Contains a `GRangesList` for peaks from the Bcells and Eryths clusters
}
\examples{
zeller_H3K4me3_peaks <- zeller_H3K4me3_peaks
}
\keyword{datasets}
