% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeller_H3K9me3.R
\docType{data}
\name{zeller_H3K9me3_metadata}
\alias{zeller_H3K9me3_metadata}
\title{H3K9me3 sortChIC metadata}
\format{
A vector of `length(ncol(zeller_H3K9me3_matrix))`
}
\usage{
zeller_H3K9me3_metadata
}
\description{
These datasets contain a subset of the cells from the H3K9me3 dataset of Zeller et al. (2023)
}
\details{
Contains a vector of the FACS-based celltype per column in `zeller_H3K9me3_matrix`.
}
\examples{
test_matrix <- zeller_H3K9me3_matrix
test_matrix_clusters <- zeller_H3K9me3_metadata
}
\keyword{datasets}
